Added Color to the map
Currently only aei_pr_soiler is visualized in NUTS Level 3. The whole
thing is sill static. Therefore the next step is to make it it dynamic.
Furthermore the current build throws warnings due to the removal of the
NaN areas from an already compiled ColumnDataSource. Should be fixable
rather easy, tho.