implemented map redraw on selection change
Note that not all datasets contain observations in all NUTS levels.
aact_ali01 contains not a single one in any of the 3 NUTS levels and
there can't be displayed, yet we generate geojsons for it and mark them
as selectable. For trng_lfse_04 we do not have data for NUTS level 3.
The most complete dataset is aei_pr_soiler at the moment.

The generated GeoJSONs are equal to the source NUTS.geojson ...